pyproject.toml
[project]
name = "flowtts-byok"
version = "1.0.0"
description = "FlowTTS BYOK - 腾讯云 FlowTTS 语音合成服务封装"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "chicogong", email = "gonghaoran@outlook.com"}
]
keywords = ["tts", "text-to-speech", "tencent-cloud", "flowtts"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "tencentcloud-sdk-python>=3.0.1200",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=3.5.0",
    "pytest>=7.4.0",
]
gradio = [
    "gradio>=3.50.0",
    "huggingface_hub>=0.19.0",
]
streamlit = [
    "streamlit>=1.28.0",
]

[project.urls]
Homepage = "https://github.com/chicogong/flowtts-byok"
Repository = "https://github.com/chicogong/flowtts-byok"
Documentation = "https://github.com/chicogong/flowtts-byok#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

# Ruff 配置
[tool.ruff]
target-version = "py310"
line-length = 88
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".cog",
    "*.ipynb",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict
]

[tool.ruff.lint.isort]
known-first-party = ["predict"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Pytest 配置
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
